объединил 2 маппера VacancyLocalMapper и FavoritesLocalMapper, чтобы уменьшить число зависимостей
