+++ app/src/main/java/ru/practicum/android/diploma/data/repository/VacancyRepositoryImpl.kt
@@ -4,7 +4,6 @@ import android.util.Log
 import retrofit2.HttpException
 import ru.practicum.android.diploma.data.local.dao.VacancyDao
 import ru.practicum.android.diploma.data.local.dao.VacancyInFavoritesDao
-import ru.practicum.android.diploma.data.local.mapper.FavoritesLocalMapper
 import ru.practicum.android.diploma.data.local.mapper.VacancyLocalMapper
 import ru.practicum.android.diploma.data.remote.api.ApiService
 import ru.practicum.android.diploma.data.remote.dto.response.ApiResponse
@@ -25,8 +24,7 @@ class VacancyRepositoryImpl(
     private val vacancyInFavoritesDao: VacancyInFavoritesDao,
     private val vacancyRemoteMapper: VacancyRemoteMapper,
     private val vacancyLocalMapper: VacancyLocalMapper,
-    private val vacancyRequestMapper: VacancyRequestMapper,
-    private val favoritesLocalMapper: FavoritesLocalMapper
+    private val vacancyRequestMapper: VacancyRequestMapper
 ) : IVacancyRepository {
 
     override suspend fun getVacancies(request: VacancySearchRequest): ApiResponse<VacancySearchResult> {
@@ -91,7 +89,7 @@ class VacancyRepositoryImpl(
     }
 
     override suspend fun addVacancyToFavorites(vacancy: Vacancy) {
-        vacancyInFavoritesDao.insertVacancy(favoritesLocalMapper.toEntity(vacancy))
+        vacancyInFavoritesDao.insertVacancy(vacancyLocalMapper.toFavoritesEntity(vacancy))
     }
 
     override suspend fun deleteVacancyFromFavorites(id: String) {
@@ -101,13 +99,13 @@ class VacancyRepositoryImpl(
     override suspend fun getFavoriteVacancies(): List<Vacancy> {
         return vacancyInFavoritesDao
             .getAll()
-            .map { favoritesLocalMapper.mapFromDb(it) }
+            .map { vacancyLocalMapper.mapFromDb(it) }
     }
 
     override suspend fun getFavoriteVacancyById(id: String): Vacancy? {
         val vacancyEntity = vacancyInFavoritesDao.getVacancyById(id)
         if (vacancyEntity == null) return null
-        return favoritesLocalMapper.mapFromDb(vacancyEntity)
+        return vacancyLocalMapper.mapFromDb(vacancyEntity)
     }
 
     override suspend fun checkIsVacancyInFavoritesById(id: String): Boolean {
