+++ app/src/main/java/ru/practicum/android/diploma/di/AppModule.kt
@@ -11,7 +11,6 @@ import retrofit2.converter.gson.GsonConverterFactory
 import ru.practicum.android.diploma.BuildConfig
 import ru.practicum.android.diploma.data.local.database.AppDatabase
 import ru.practicum.android.diploma.data.local.mapper.AreaLocalMapper
-import ru.practicum.android.diploma.data.local.mapper.FavoritesLocalMapper
 import ru.practicum.android.diploma.data.local.mapper.IndustryLocalMapper
 import ru.practicum.android.diploma.data.local.mapper.VacancyLocalMapper
 import ru.practicum.android.diploma.data.remote.api.ApiService
@@ -66,7 +65,6 @@ val appModule = module {
     single { IndustryRemoteMapper() }
     single { VacancyRemoteMapper(get(), get()) }
     single { VacancyRequestMapper() }
-    single { FavoritesLocalMapper() }
 
     // Local Mappers
     single { AreaLocalMapper() }
@@ -109,7 +107,7 @@ val appModule = module {
     // Repositories
     single<IAreaRepository> { AreaRepositoryImpl(get(), get(), get(), get()) }
     single<IIndustryRepository> { IndustryRepositoryImpl(get(), get(), get(), get()) }
-    single<IVacancyRepository> { VacancyRepositoryImpl(get(), get(), get(), get(), get(), get(), get()) }
+    single<IVacancyRepository> { VacancyRepositoryImpl(get(), get(), get(), get(), get(), get()) }
 
     // Use Cases
     single { SearchVacanciesUseCase(get()) }
